Moved unit and integration tests into separate folders
