+++ tests/spec/spec_helper.rb
@@ -1,0 +1,26 @@
+# encoding: UTF-8
+require 'rubygems'
+require 'bundler/setup'
+require 'winrm'
+require 'json'
+require_relative '../matchers'
+
+module StubHelper
+  def stubbed_response(file)
+    File.read("tests/spec/stubs/responses/#{file}")
+  end
+end
+
+# Strip leading whitespace from each line that is the same as the
+# amount of whitespace on the first line of the string.
+# Leaves _additional_ indentation on later lines intact.
+# and remove newlines.
+class String
+  def unindent
+    gsub(/^#{self[/\A[ \t]*/]}/, '').gsub("\n", '')
+  end
+end
+
+RSpec.configure do |config|
+  config.include(StubHelper)
+end
