+++ tests/spec/stubs/responses/get_command_output_response.xml.erb
