+++ tests/spec/stubs/responses/open_shell_v1.xml
