+++ tests/spec/stubs/responses/open_shell_v2.xml
