+++ tests/spec/stubs/responses/soap_fault_v1.xml
