+++ tests/spec/stubs/responses/soap_fault_v2.xml
