+++ tests/spec/stubs/responses/wmi_error_v2.xml
